server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
redis==5.0.1
rq==1.15.1
boto3==1.33.6
httpx==0.25.2
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
